Support AUD & ISS claims, all data into token
This commit adds support for the AUD (Audience) and ISS (Issuer) JWT
claims. See jwt.io and the relevant RFC for details.
Update the readme to show these supported claims and setting them via
env vars.
General cleanup/tidy of readme as a whole.
The service was returning the user model as json, with some information
being at the top level of the response as well as within the token. Now
the service just returns the token with all of the information inside of
it.